Rename .termType to .valueType
Per @bergos - The Term can be seen as a container. `.termType` would
mean the type of the Term, so the interface name/type. But it
describes the type of the value ("iri", "bnode", "literal", and
"variable"). We had a very long discussion about these values. So
let's keep these values and rename `.termType` to `.valueType`. Than
it would describe the type of the value the container holds.